got it up and running but im unsure if this is the way

+++ src/Imaginengion/Core/UUID.zig
@@ -1,5 +1,5 @@
 const std = @import("std");
-pub fn GenUUID() !u128 {
+pub fn GenUUID() !u64 {
     var prng = std.Random.DefaultPrng.init(blk: {
         var seed: u64 = undefined;
         try std.posix.getrandom(std.mem.asBytes(&seed));
@@ -7,5 +7,5 @@ pub fn GenUUID() !u128 {
     });
     const rand = prng.random();
 
-    return rand.uintAtMost(u128, ~@as(u128, 0) - 1);
+    return rand.uintAtMost(u64, ~@as(u64, 0) - 1);
 }
